Add files to separate tests
